Make dedicated packages for adapter implementations
