+++ pkg/operations/imported/adapter.go
@@ -1,0 +1,169 @@
+package imported
+
+import (
+	"fmt"
+
+	"github.com/rancher/rancher/pkg/apis/management.cattle.io/v3"
+	mgmtv3 "github.com/rancher/rancher/pkg/apis/management.cattle.io/v3"
+	"github.com/rancher/rancher/pkg/capr"
+	"github.com/rancher/rancher/pkg/operations"
+	. "github.com/rancher/rancher/pkg/operations"
+	"github.com/rancher/rancher/pkg/plan"
+	"github.com/rancher/rancher/pkg/wrangler"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/runtime"
+)
+
+type ImportedAdapter struct {
+	cluster *v3.Cluster
+	clients *wrangler.CAPIContext
+}
+
+// WaitForRegister waits for all machine-plan secrets to be created, ensuring the system-agent has checked in for
+// all expected nodes.
+// All machine-plans secrets are listed and compared to the count of mgmtv3.Node objects for the cluster.
+func (a *ImportedAdapter) WaitForRegister() (bool, error) {
+	secretList, err := a.clients.Core.Secret().List(a.cluster.Name, metav1.ListOptions{
+		LabelSelector: fmt.Sprintf("%s=%s", capr.ClusterNameLabel, a.cluster.Name),
+		FieldSelector: fmt.Sprintf("type=%s", capr.SecretTypeMachinePlan),
+	})
+	if err != nil {
+		return false, err
+	}
+
+	secrets := secretList.Items
+
+	machines, err := a.clients.Mgmt.Node().Cache().List(a.cluster.Name, labels.Everything())
+	if err != nil {
+		return false, err
+	}
+
+	// If the counts don't match upfront, we already know it's not a 1:1 match
+	if len(secrets) != len(machines) {
+		return false, nil
+	}
+
+	// Track the names of the machines we expect to see
+	// Using a map[string]bool to easily check existence and prevent duplicates
+	expectedMachines := make(map[string]bool, len(machines))
+	for _, machine := range machines {
+		expectedMachines[machine.Name] = true
+	}
+
+	// Verify that every secret maps to a unique expected machine
+	for _, secret := range secrets {
+		if secret.Labels == nil {
+			return false, nil
+		}
+
+		machineName, exists := secret.Labels[capr.MachineNameLabel]
+
+		// If the label is missing, or it maps to a machine name we haven't seen/already matched
+		if !exists || !expectedMachines[machineName] {
+			return false, nil
+		}
+
+		// Mark this machine as "matched" by deleting it from the expected map.
+		// This naturally catches duplicate secrets pointing to the same machine.
+		delete(expectedMachines, machineName)
+	}
+
+	// If the map is empty, we have a perfect, duplicate-free 1:1 match
+	return len(expectedMachines) == 0, nil
+}
+
+// RuntimeCommand returns the command used to interact with the distro CLI (RKe2/K3s).
+func (a *ImportedAdapter) RuntimeCommand() string {
+	if a.cluster.Status.Provider == "rke2" {
+		return "rke2"
+	}
+	return "k3s"
+}
+
+// ServerUnit returns the systemd unit name for a distro server node.
+func (a *ImportedAdapter) ServerUnit() string {
+	if a.cluster.Status.Provider == "rke2" {
+		return "rke2-server"
+	}
+	return "k3s"
+}
+
+// RenderProbes renders the probes for a given machine-plan secret based on its role.
+// Currently custom data directories, probes, and using ipv4 as the primary ip family are not supported.
+func (a *ImportedAdapter) RenderProbes(secret *corev1.Secret) (map[string]plan.Probe, error) {
+	var (
+		runtime    = a.RuntimeCommand()
+		probeNames []string
+		probes     = map[string]plan.Probe{}
+	)
+
+	if runtime != capr.RuntimeK3S && IsEtcd(secret) {
+		probeNames = append(probeNames, ETCDProbeName)
+	}
+	if IsControlPlane(secret) {
+		probeNames = append(probeNames, KubeAPIServerProbeName)
+		probeNames = append(probeNames, KubeControllerManagerProbeName)
+		probeNames = append(probeNames, KubeSchedulerProbeName)
+	}
+	if !(And(IsEtcd, Not(IsControlPlane))(secret) && runtime == capr.RuntimeK3S) {
+		// k3s doesn't run the kubelet on etcd only nodes
+		probeNames = append(probeNames, KubeletProbeName)
+	}
+
+	for _, probeName := range probeNames {
+		probes[probeName] = AllProbes[probeName]
+	}
+
+	dataDir := "/var/lib/rancher/rke2"
+	if runtime == capr.RuntimeK3S {
+		dataDir = "/var/lib/rancher/k3s"
+	}
+
+	probes = InsertDataDirForProbes(dataDir, probes)
+
+	// only support ipv4, need to implement per-node extraction mechanism
+	loopbackAddress := "127.0.0.1"
+
+	if IsControlPlane(secret) {
+		kcmProbe, err := renderSecureProbe(probes[KubeControllerManagerProbeName], loopbackAddress, DefaultKubeControllerManagerPort, DefaultKubeControllerManagerCertDir, DefaultKubeControllerManagerCert)
+		if err != nil {
+			return probes, err
+		}
+		probes[KubeControllerManagerProbeName] = kcmProbe
+
+		ksProbe, err := renderSecureProbe(probes[KubeSchedulerProbeName], loopbackAddress, DefaultKubeSchedulerPort, DefaultKubeSchedulerCertDir, DefaultKubeSchedulerCert)
+		if err != nil {
+			return probes, err
+		}
+		probes[KubeSchedulerProbeName] = ksProbe
+	}
+
+	probes = ReplaceURLForProbes(probes, loopbackAddress)
+
+	return probes, nil
+}
+
+// renderSecureProbe takes the existing argument value and renders a secure probe using the argument values and an error
+// if one occurred.
+func renderSecureProbe(probe plan.Probe, loopbackAddress, securePort string, certDir string, cert string) (plan.Probe, error) {
+	TLSCert := certDir + "/" + cert
+	return ReplaceCACertAndPortForProbes(probe, TLSCert, loopbackAddress, securePort)
+}
+
+func init() {
+	operations.RegisterAdapter(mgmtv3.SchemeGroupVersion.WithKind("Cluster"), func(clients *wrangler.CAPIContext, unstructured *unstructured.Unstructured) (operations.Adapter, error) {
+		var cluster *mgmtv3.Cluster
+		err := runtime.DefaultUnstructuredConverter.FromUnstructured(unstructured.Object, &cluster)
+		if err != nil {
+			return nil, err
+		}
+
+		return &ImportedAdapter{
+			cluster: cluster,
+			clients: clients,
+		}, nil
+	})
+}
